Move components back to toplevel
I forgot this was more standard for Next.js so we just keep it that way.
This means we need to change less to standard postcss configs.
