+++ lib/talks.js
