+++ pages/_app.js
@@ -1,5 +1,5 @@
-import Layout from "../lib/components/layout";
-import '../lib/components/layout.css';
+import Layout from "../components/layout";
+import '../components/layout.css';
 
 export default function App({Component, pageProps}) {
   return (
